+++ Tests/Obfy.ScenarioTests/Fixtures/UnitySample/UnityEngine.cs
@@ -1,0 +1,15 @@
+namespace UnityEngine;
+
+public class Object
+{
+    public string name = "";
+}
+
+public class GameObject : Object
+{
+}
+
+public class MonoBehaviour
+{
+    public GameObject gameObject { get; set; } = new();
+}